// packages/shared/src/validators/company.d.ts
import { z } from "zod";
export declare const createCompanySchema: z.ZodObject<{
    name: z.ZodString;
    description: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    budgetMonthlyCents: z.ZodDefault<z.ZodOptional<z.ZodNumber>>;
}, "strip", z.ZodTypeAny, {
    name: string;
    budgetMonthlyCents: number;
    description?: string | null | undefined;
}, {
    name: string;
    description?: string | null | undefined;
    budgetMonthlyCents?: number | undefined;
}>;
export type CreateCompany = z.infer<typeof createCompanySchema>;
export declare const updateCompanySchema: z.ZodObject<{
    name: z.ZodOptional<z.ZodString>;
    description: z.ZodOptional<z.ZodNullable<z.ZodOptional<z.ZodString>>>;
    budgetMonthlyCents: z.ZodOptional<z.ZodDefault<z.ZodOptional<z.ZodNumber>>>;
} & {
    status: z.ZodOptional<z.ZodEnum<["active", "paused", "archived"]>>;
    spentMonthlyCents: z.ZodOptional<z.ZodNumber>;
    requireBoardApprovalForNewAgents: z.ZodOptional<z.ZodBoolean>;
    brandColor: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    logoAssetId: z.ZodOptional<z.ZodNullable<z.ZodString>>;
}, "strip", z.ZodTypeAny, {
    name?: string | undefined;
    description?: string | null | undefined;
    status?: "active" | "paused" | "archived" | undefined;
    budgetMonthlyCents?: number | undefined;
    spentMonthlyCents?: number | undefined;
    requireBoardApprovalForNewAgents?: boolean | undefined;
    brandColor?: string | null | undefined;
    logoAssetId?: string | null | undefined;
}, {
    name?: string | undefined;
    description?: string | null | undefined;
    status?: "active" | "paused" | "archived" | undefined;
    budgetMonthlyCents?: number | undefined;
    spentMonthlyCents?: number | undefined;
    requireBoardApprovalForNewAgents?: boolean | undefined;
    brandColor?: string | null | undefined;
    logoAssetId?: string | null | undefined;
}>;
export type UpdateCompany = z.infer<typeof updateCompanySchema>;
export declare const updateCompanyBrandingSchema: z.ZodEffects<z.ZodObject<{
    name: z.ZodOptional<z.ZodString>;
    description: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    brandColor: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    logoAssetId: z.ZodOptional<z.ZodNullable<z.ZodString>>;
}, "strict", z.ZodTypeAny, {
    name?: string | undefined;
    description?: string | null | undefined;
    brandColor?: string | null | undefined;
    logoAssetId?: string | null | undefined;
}, {
    name?: string | undefined;
    description?: string | null | undefined;
    brandColor?: string | null | undefined;
    logoAssetId?: string | null | undefined;
}>, {
    name?: string | undefined;
    description?: string | null | undefined;
    brandColor?: string | null | undefined;
    logoAssetId?: string | null | undefined;
}, {
    name?: string | undefined;
    description?: string | null | undefined;
    brandColor?: string | null | undefined;
    logoAssetId?: string | null | undefined;
}>;
export type UpdateCompanyBranding = z.infer<typeof updateCompanyBrandingSchema>;
//# sourceMappingURL=company.d.ts.map